Inital refactoring
Each class in its own file, Logger as interface
