+++ AzureDevOps.WikiPDFExport/ConsoleLogger.cs
@@ -1,0 +1,42 @@
+﻿using Microsoft.Extensions.Logging;
+using System;
+
+namespace azuredevops_export_wiki
+{
+    internal class ConsoleLogger : ILogger
+    {
+        private Options _options;
+
+        internal ConsoleLogger(Options options) { _options = options; }
+
+        public void Log(string msg, LogLevel logLevel = LogLevel.Information, int indent = 0)
+        {
+            var indentString = new string(' ', indent * 2);
+            if (_options.Debug && logLevel == LogLevel.Debug)
+            {
+                Console.WriteLine(indentString + msg);
+            }
+
+            if (_options.Verbose && logLevel == LogLevel.Information)
+            {
+                Console.WriteLine(indentString + msg);
+            }
+
+            if (logLevel == LogLevel.Warning)
+            {
+                var color = Console.ForegroundColor;
+                Console.ForegroundColor = ConsoleColor.Yellow;
+                Console.WriteLine(indentString + $"WARN: {msg}");
+                Console.ForegroundColor = color;
+            }
+
+            if (logLevel == LogLevel.Error)
+            {
+                var color = Console.ForegroundColor;
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine(indentString + $"ERR: {msg}");
+                Console.ForegroundColor = color;
+            }
+        }
+    }
+}
